drivers: memc: memc_mcux_flexspi: fix handling of multiple devices
Fix handling of multiple devices for MEMC FlexSPI driver. The
FlexSPI_UpdateLUT function expects that LUT array sizes be given
as a count of 4 byte entries, but the FlexSPI IP expects each sequence
to be 16 bytes in length.

To resolve this, update the definition of MEMC_FLEXSPI_CMD_SIZE. This
constant is used by all in tree drivers to determine the count argument
for memc_flexspi_set_device_config, so each driver should now pass the
correct count of LUT command sequences.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>