boards: arm: mimxrt685_evk_cm33: add PSRAM definition
Add PSRAM definition for RT685 EVK. The PSRAM can theoretically be
clocked at 200MHz, but the Octal Flash on this EVK has much lower clock
frequency restrictions of 50MHz due to not having a DQS pin connected
for read strobe. Therefore, the PSRAM is also kept to a lower clock
frequency.

Required FlexSPI peripheral configuration is also added, since this was
previously not needed to interface with the Octal Flash when running in
XIP mode.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>